feat: refactor structure and get pagination right
